backend/requirements.txt
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
dnspython==2.4.2
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
fonttools==4.47.2
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
mysql-connector-python==8.3.0
numpy==1.26.3
packaging==23.2
pillow==10.2.0
pymongo==4.6.1
pyparsing==3.1.1
python-dateutil==2.8.2
six==1.16.0
Werkzeug==3.0.1
paho-mqtt>=1.5.0
